workflows: build_and_deploy: use a single prefix for test jobs
Test jobs used to be prefixed by either `testbot` or `qemu` to represent
either physical or virtual environments.

This change uses a customizable unique job prefix that defaults to `test`
that will be used to define whether to release as final, abstracting the
actual run environment from the decision.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>